Fix more pointer to integer size mismatches in WIN64/LLP64
1) struct NSFastEnumerationState.state must also be 64 bits wide to store 64 bit pointers.
2) Fix minor declaration mistakes, and replace bzero(3) with memset(3) as bzero(3) is deprecated and not portable.

+++ source/CHDoublyLinkedList.m
@@ -255,7 +255,7 @@ static size_t kCHDoublyLinkedListNodeSize = sizeof(CHDoublyLinkedListNode);
 	if (currentNode == tail)
 		state->state = 1; // used as a termination flag
 	else
-		state->state = (unsigned long)currentNode;
+		state->state = (uintptr_t) currentNode;
     return batchCount;
 }
